requirements.txt
requests
supabase
python-dotenv
streamlit
plotly
pandas
apscheduler
streamlit-autorefresh
psycopg2-binary